requirements.txt
sqlalchemy==2.0.40
langchain==0.3.25
langchain-community==0.3.25
langchain-groq==0.2.4
fastapi==0.135.3
uvicorn[standard]==0.42.0
python-dotenv==1.0.1
pandas==2.2.3
rapidfuzz==3.11.0
numpy==2.2.4
slowapi==0.1.9
